refines orgunit invoice selection page, implements more dynamic selection of invoices
